Move recordConsoleLogs into test directory, add/update doc comments
