+++ lib/src/util/console_log_utils.dart
@@ -20,40 +20,37 @@ import 'dart:js';
 import 'package:meta/meta.dart';
 import 'package:react/react_client/react_interop.dart';
 
-/// Runs a provided callback and returns the logs that occur during the runtime
+/// Runs a provided [callback] and [print]s each log that occurs during the runtime
 /// of that function.
 ///
-/// Can be used to capture logs, warnings, or errors as specified by setting
-/// the [configuration]. To set the [configuration], pass in the corresponding
-/// config class ([ConsoleConfig.log], [ConsoleConfig.warn], [ConsoleConfig.error], [ConsoleConfig.all]).
+/// [print] is called in the same Zone as this function was called in, similar to a Stream subscription.
+///
+/// Can be used to print logs, warnings, or errors based on [configuration].
 ///
 /// The function assumes that any `propType` warnings that occur during
 /// the function runtime should be captured. Consequently, the `PropType` cache
 /// is reset prior to calling the provided callback.
-///
-/// If any errors are thrown during the callback, e.g. during a render that expects
-/// props that are not valid, the errors will be caught to allow the test to complete.
-List<String> recordConsoleLogs(
-  Function() callback, {
-  ConsoleConfig configuration = ConsoleConfig.all,
-}) {
-  final consoleLogs = <String>[];
-
-  final stopSpying = startSpyingOnConsoleLogs(configuration: configuration, onLog: consoleLogs.add);
-  try {
-    callback();
-  } finally {
-    stopSpying();
-  }
-
-  return consoleLogs;
-}
-
 T printConsoleLogs<T>(
   T Function() callback, {
   ConsoleConfig configuration = ConsoleConfig.all,
+}) => spyOnConsoleLogs(callback, configuration: configuration, onLog: print);
+
+/// Runs a provided [callback] and calls [onLog] for each log that occurs during the runtime
+/// of that function.
+///
+/// [onLog] is called in the same Zone as this function was called in, similar to a Stream subscription.
+///
+/// Can be used to capture logs, warnings, or errors based on [configuration].
+///
+/// The function assumes that any `propType` warnings that occur during
+/// the function runtime should be captured. Consequently, the `PropType` cache
+/// is reset prior to calling the provided callback.
+T spyOnConsoleLogs<T>(
+  T Function() callback, {
+  @required void Function(String) onLog,
+  ConsoleConfig configuration = ConsoleConfig.all,
 }) {
-  final stopSpying = startSpyingOnConsoleLogs(configuration: configuration, onLog: print);
+  final stopSpying = startSpyingOnConsoleLogs(configuration: configuration, onLog: onLog);
   try {
     return callback();
   } finally {
@@ -61,6 +58,16 @@ T printConsoleLogs<T>(
   }
 }
 
+/// Starts spying on concole logs, calling [onLog] for each log that occurs until the
+/// returned function (`stopSpying`) is called.
+///
+/// [onLog] is called in the same Zone as this function was called in, similar to a Stream subscription.
+///
+/// Can be used to capture logs, warnings, or errors based on [configuration].
+///
+/// The function assumes that any `propType` warnings that occur during
+/// the function runtime should be captured. Consequently, the `PropType` cache
+/// is reset prior to calling the provided callback.
 void Function() startSpyingOnConsoleLogs({
   ConsoleConfig configuration = ConsoleConfig.all,
   @required void Function(String) onLog,
@@ -99,30 +106,33 @@ void _resetPropTypeWarningCache() {
   } catch (_) {}
 }
 
-/// Configuration class that sets options within [recordConsoleLogs].
+/// Configuration class that sets options within console-log-recording functions:
+///
+/// - [printConsoleLogs]
+/// - [spyOnConsoleLogs]
+/// - [startSpyingOnConsoleLogs]
 @sealed
 class ConsoleConfig {
   const ConsoleConfig._(this.logType);
 
-  /// The type of log to capture while running the callbacks within
-  /// [recordConsoleLogs].
+  /// The type of log to capture, corresponding to a method on the JS `console` object,
+  /// or `'all'` for all types.
   ///
-  /// Must be `'warn'`, `'error'`, `'log'` or `'all'`.
+  /// Must be either `'all'` one of the values within [types].
   final String logType;
 
   /// The possible console types that have different log contexts.
   static const Set<String> types = {'error', 'log', 'warn'};
 
-  /// The configuration needed to capture logs while running [recordConsoleLogs].
+  /// Captures calls to `console.log`.
   static const ConsoleConfig log = ConsoleConfig._('log');
 
-  /// The configuration needed to capture warnings while running [recordConsoleLogs].
+  /// Captures calls to `console.warn`.
   static const ConsoleConfig warn = ConsoleConfig._('warn');
 
-  /// The configuration needed to capture errors while running [recordConsoleLogs].
+  /// Captures calls to `console.error`.
   static const ConsoleConfig error = ConsoleConfig._('error');
 
-  /// The configuration that will capture all logs, whether they be logs, warnings,
-  /// or errors.
+  /// Captures calls to `console.log`, `console.warn`, and `console.error`.
   static const ConsoleConfig all = ConsoleConfig._('all');
 }
